Add background image
I added an image to the project that will be used as the background for the web app.

I decided to choose Stonehenge since I'm an archaeologist. Besides I have never been there, and is one of my to go sites in 2024. Also, everyone has at least seen a picture of it, so it should be easy for the user to relate with it.