Add components - Login
Added the login.css with the styling of the elements created in the login.js.

Combined with the existing code it should create a web app that has the Stonehenge image behind. In the centre there should be a rectangle (semi-transparent) with a title (Welcome to Prehistory!) and two input spaces for the user to write the username and password.