Create the Welcome page
After the user is logged in correctly, it will enter a new page, called welcome page. In here it will see a greetings message, with the user name on it, some facts about Stonehenge and an image.
Besides I also included a logout button that will take the user back to the main page.