Uploading screenshots and video
Uploading the screenshots and video that prove the app is working correctly.
In the video I input different passwords, the same password and user name in uppercase. This way it is possible to see that it only unlock when the correct combination is inserted.
After this it will lead to the welcome page. If there is an error, it will lead to the error page.